+++ src/sc_keeper/api.py
@@ -1025,11 +1025,17 @@ def search_databases(
     engine_version: options.database_engine_version = None,
     vcpus_min: options.vcpus_min = 1,
     vcpus_max: options.vcpus_max = None,
+    architecture: options.architecture = None,
+    cpu_allocation: options.cpu_allocation = None,
     memory_min: options.memory_min = None,
+    network_speed_baseline_min: options.network_speed_baseline_min = None,
+    network_speed_max_min: options.network_speed_max_min = None,
+    network_storage_speed_baseline_min: options.network_storage_speed_baseline_min = None,
+    network_storage_speed_max_min: options.network_storage_speed_max_min = None,
     benchmark_id: options.benchmark_id = "pgbench:heavy_read_only",
     benchmark_config: options.benchmark_config = None,
-    benchmark_score_min: options.benchmark_score_min = None,
-    benchmark_score_per_price_min: options.benchmark_score_per_price_min = None,
+    benchmark_score_min: options.database_benchmark_score_min = None,
+    benchmark_score_per_price_min: options.database_benchmark_score_per_price_min = None,
     ha: options.database_ha = None,
     ha_strategy: options.database_ha_strategy = None,
     max_read_replicas_min: options.database_max_read_replicas_min = None,
@@ -1149,8 +1155,24 @@ def search_databases(
         conditions.add(Database.vcpus >= vcpus_min)
     if vcpus_max:
         conditions.add(Database.vcpus <= vcpus_max)
+    if architecture:
+        conditions.add(Server.cpu_architecture.in_(architecture))
+    if cpu_allocation:
+        conditions.add(Server.cpu_allocation.in_(cpu_allocation))
     if memory_min:
         conditions.add(Database.memory_amount >= memory_min * 1024)
+    if network_speed_baseline_min:
+        conditions.add(Server.network_speed_baseline >= network_speed_baseline_min)
+    if network_speed_max_min:
+        conditions.add(Server.network_speed_max >= network_speed_max_min)
+    if network_storage_speed_baseline_min:
+        conditions.add(
+            Server.network_storage_speed_baseline >= network_storage_speed_baseline_min
+        )
+    if network_storage_speed_max_min:
+        conditions.add(
+            Server.network_storage_speed_max >= network_storage_speed_max_min
+        )
     if ha:
         jh = func.json_each(Database.ha).table_valued("value")
         ha_count = (
@@ -1259,8 +1281,23 @@ def search_databases(
         "selected_benchmark_score_per_price": _live_price_order_min_price,
     }
 
+    needs_server_join = bool(
+        architecture
+        or cpu_allocation
+        or network_speed_baseline_min
+        or network_speed_max_min
+        or network_storage_speed_baseline_min
+        or network_storage_speed_max_min
+    )
+
     if add_total_count_header:
         query = select(func.count()).select_from(Database)
+        if needs_server_join:
+            query = query.join(
+                Server,
+                (Database.vendor_id == Server.vendor_id)
+                & (Database.server_id == Server.server_id),
+            )
         if (
             only_active
             or only_orderable
@@ -1328,6 +1365,12 @@ def search_databases(
 
     query = select(*select_cols)
     query = query.join(Database.vendor)
+    if needs_server_join:
+        query = query.join(
+            Server,
+            (Database.vendor_id == Server.vendor_id)
+            & (Database.server_id == Server.server_id),
+        )
     query = query.join(
         DatabaseExtra,
         (Database.vendor_id == DatabaseExtra.vendor_id)