+++ src/sc_keeper/parameters.py
@@ -1069,3 +1069,32 @@ database_sla_min = Annotated[
         },
     ),
 ]
+
+database_benchmark_score_min = Annotated[
+    Optional[float],
+    Query(
+        title="Required benchmark score",
+        description=(
+            "Required value of the selected benchmark score "
+            "(default benchmark_id=pgbench:heavy_read_only)."
+        ),
+        json_schema_extra={
+            "category_id": DatabaseFilterCategory.PERFORMANCE,
+        },
+    ),
+]
+
+database_benchmark_score_per_price_min = Annotated[
+    Optional[float],
+    Query(
+        title="Required benchmark score/price",
+        description=(
+            "Required value of the selected benchmark score per USD/hr "
+            "(default benchmark_id=pgbench:heavy_read_only, using the best on-demand price)."
+        ),
+        json_schema_extra={
+            "category_id": DatabaseFilterCategory.PERFORMANCE,
+            "unit": "/USD",
+        },
+    ),
+]